[Enhancement] Filetype Associations (#81)
This pull request adds the feature for filetype associations.

Most of the code is brought over from nx-hbmenu. I have tested the config I have for LÖVE Potion and it works. Icons are used from tex3ds generated t3x files which must be RGB565 and 48x48 (as usual for icon sizes). Some of the code may be a bit all over the place, but I tried to keep it organized with everything else.

This will be super useful for other projects as well.