Make README more beginner friendly
The Homebrew Launcher is most likely the most popular homebrew application
for the Nintendo 3DS.

Many years ago, it was a great starting point for me before attempting to
write my own "hello world" type of application, but I kept getting stuck.
I now know how to contribute back, so this is my attempt at authoring the
things I wish I would have read back then.

- Use GitHub-styled admonition boxes in the introduction and in the build
  instructions
- Include some resources explaining where the person wanting to build
  the launcher by themselves would have to go if they are just starting
  out. The Homebrew Launcher was a personal starting point back then, and
  I would presume that this would be the case for others too.
- Environment variables can be an issue if you are not using specific kinds
  of systems that seem to be supported by devkitPro.

In the case of `make` and environment variables, I explained some things that
would seem pretty self-explanatory to most developers, but there are (and were)
tinkerers that can be very young. In the cases of chose to explain the technical
terms, some details and the theory behind them. It contains enough pointers for
the reader to look up what they need for their own setup online.