Prioritize .3dsx files over folders in menu sorting
This change deviates away from a de facto standard of putting the
directories first, and showing the files second. This behavior can be
observed in a variety of different programs, such as the ls utility
or GodMode9; however, it does not work that well here.

In the real world, folders seem to be mostly used for application data
and do not contain anything that can be launched. If the user puts in
enough effort to organize apps in folders, however, I would imagine
that the user would probably use the "Star" functionality if absolutely
necessary and would most likely maintain the `3ds` folder "clean".